refactor: rename submodule to ci_generation
