+++ src/generator/ci_generation/templates/azure_pipelines.rs
