+++ src/generator/ci_generation/templates/cloud_build.rs
