+++ src/generator/ci_generation/templates/github_actions.rs
